% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{create_tags}
\alias{create_tags}
\title{Create a new tags container}
\usage{
create_tags(filename, prefix = "", default = TRUE)
}
\arguments{
\item{filename}{the name of the file where the tags will be stored.}

\item{prefix}{the string to prepend to all tag names.}

\item{default}{if `TRUE` make this the default tags container.}
}
\description{
It creates a new tags container that will be associated with the given
`filename` and the tags will be prefixed with `prefix` string.
}
