% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt}
\alias{fmt}
\title{Formats given value.}
\usage{
fmt(x, prefix = "", suffix = "", ...)
}
\arguments{
\item{x}{the value to be formatted}

\item{prefix}{the string to be prepended}

\item{suffix}{the string to be appended}

\item{...}{parameters to be passed to the actual formatter. The number
formatter accepts `floor` and `ceiling` to indicate that given value
should be rounded down or up to a full integer before formatting or
`digits` to indicate the number of digits the number should be rounded
to.}
}
\description{
The `x` value is formatted according to its class with an additional
`prefix` and `suffix` prepended and appended respectively.
}
