% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{oom}
\alias{oom}
\title{Scale the value to its closest order of magnitude}
\usage{
oom(x, suffixes = c("", "K", "M", "G", "T"))
}
\arguments{
\item{x}{the value to scale}

\item{suffixes}{are the suffixes to be add. Each corresponds to the exponent
of the base.}
}
\description{
It uses the base of 1000.
}
