% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{percent}
\alias{percent}
\title{Converts given value to a percent representation.}
\usage{
percent(x)
}
\arguments{
\item{x}{the value to convert}
}
\description{
It multiples the value by 100 and adds the `%` suffix attribute.
}
