% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{size}
\alias{size}
\title{Scale the value to its closest order of magnitude in the size of bytes.}
\usage{
size(x, units = c("B", "kB", "MB", "GB", "TB"))
}
\arguments{
\item{x}{the value to scale}

\item{units}{are the suffixes to be add. Each corresponds to the exponent
of the base.}
}
\description{
It uses the base of 1024.
}
