% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag}
\alias{tag}
\title{Create a new tag}
\usage{
tag(name, value, tags = get_default_tags())
}
\arguments{
\item{name}{the tag name.}

\item{value}{the tag value.}

\item{tags}{the tag container, using the default container by default.}
}
\description{
Creates a new tag named `name` with value `value` in the `tags` tags
container.
}
